requirement.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
Markdown==3.3.6
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
